Refactoring: extract functions into link_utils.rs, move get_dna() and get_wasm() into context.
