+++ core/src/context.rs
@@ -1,15 +1,19 @@
 use action::ActionWrapper;
 use holochain_core_types::{
     cas::storage::ContentAddressableStorage, eav::EntityAttributeValueStorage, entry::agent::Agent,
+    dna::{Dna, wasm::DnaWasm},
     error::HolochainError,
 };
 use instance::Observer;
 use logger::Logger;
 use persister::Persister;
 use state::State;
-use std::sync::{
-    mpsc::{sync_channel, SyncSender},
-    Arc, Mutex, RwLock, RwLockReadGuard,
+use std::{
+    sync::{
+        mpsc::{sync_channel, SyncSender},
+        Arc, Mutex, RwLock, RwLockReadGuard,
+    },
+    thread::sleep, time::Duration,
 };
 
 /// Context holds the components that parts of a Holochain instance need in order to operate.
@@ -90,6 +94,43 @@ impl Context {
             None => None,
             Some(ref s) => Some(s.read().unwrap()),
         }
+    }
+
+    pub fn get_dna(&self) -> Option<Dna> {
+        // In the case of genesis we encounter race conditions with regards to setting the DNA.
+        // Genesis gets called asynchronously right after dispatching an action that sets the DNA in
+        // the state, which can result in this code being executed first.
+        // But we can't run anything if there is no DNA which holds the WASM, so we have to wait here.
+        // TODO: use a future here
+        let mut dna = None;
+        let mut done = false;
+        let mut tries = 0;
+        while !done {
+            {
+                let state = self
+                    .state()
+                    .expect("Callback called without application state!");
+                dna = state.nucleus().dna();
+            }
+            match dna {
+                Some(_) => done = true,
+                None => {
+                    if tries > 10 {
+                        done = true;
+                    } else {
+                        sleep(Duration::from_millis(10));
+                        tries += 1;
+                    }
+                }
+            }
+        }
+        dna
+    }
+
+    pub fn get_wasm(&self, zome: &str) -> Option<DnaWasm> {
+        let dna = self.get_dna().expect("Callback called without DNA set!");
+        dna.get_wasm_from_zome_name(zome)
+            .and_then(|wasm| Some(wasm.clone()).filter(|_| !wasm.code.is_empty()))
     }
 }
